moviedb-318 Display matching movies from IMDB in EditMovieGUI window.
Alterations to the notes field are now correctly intercepted using the new prototype method, _InputZone.text_modified.
Also corrected was a bug in the font specification. A tk.Text object should have a default font established by the font argument to tk.Text() and not via a tag. This was causing appended text to be displayed as TkDefultFont and not the desired TkTextFont.