moviedb-318 Display matching movies from IMDB in EditMovieGUI window.
Further simplification of the ttk.Entry and Tk.Text field management has added these features to the renamed patterns module:
current value of widget is now a property.
A new `changed` method compares original and current values and returns a boolean.
The widget is instantiated within the facade allowing for all the observer mechanics to be contained within the facade class.